+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v4.0.2-1.19.2] - 2022-10-16
+### Changed
+- Use a more generic hook for catching piglin barter results to be able to support mods such as [Piglin Proliferation]
+
 ## [v4.0.1-1.19.2] - 2022-10-15
 ### Fixed
 - Fixed gui overlays from other mods rendering directly on the bartering station's interface instead of next to it
@@ -11,3 +15,4 @@ The format is based on [Keep a Changelog].
 - Initial release
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
+[Piglin Proliferation]: https://www.curseforge.com/minecraft/mc-mods/piglin-proliferation